Get PLOS metadata based on a species list

Get PLOS metadata based on a species list

sp_plos_meta(x, progress = TRUE, ...)

Arguments

x
(character/list) A vector or list of character strings to search on

progress
(logical) print a progress bar. default: TRUE

...
curl options passed on to httr::GET()

Defaults

A set of defaults are used, based on what we think are very reasonable assumptions about what most people want:

Full research articles - we assume most people want full articles, not subsets, comments, etc. - toggle this by xxx

Fields: we fetch the fields id, title, authors, but you can specify which you'd like back by xxx

Internals

Internally, this function loops over your inputs vector list passed in to the parameter x. With each loop, we use rplos::searchplos() with above defaults.

Examples

if (FALSE) {
geom <- 'POLYGON((-124.07 41.48,-119.99 41.48,-119.99 35.57,-124.07 35.57,-124.07 41.48))'
res <- sp_occ_gbif(geometry = geom, limit = 300)

# get species list first, then pass to sp_plos_meta
res %>% sp_list() %>% sp_plos_meta()

# or, pass directly to sp_plos_meta, and species list is extracted automatically
res %>% sp_plos_meta()

# combine all into a data.frame
as_df(z$`allium amplectens`)
df <- as_df(z)
names(df)
df$title
}